requirements.txt
# Flask Web Framework
Flask>=2.3.0
Flask-Login>=0.6.0
Flask-CORS>=4.0.0

# Core AI dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-google-genai>=0.0.7
google-generativeai>=0.8.0

# Document processing
PyPDF2>=3.0.0
pypdf>=3.0.0

# Vector store
chromadb>=0.4.0

# Utilities
Werkzeug>=2.3.0
toml>=0.10.0
requests>=2.28.0

# Additional typing support
typing-extensions>=4.5.0